ai_api/requirements.txt
flask
pdfplumber
python-docx
spacy
